Fixed camera looking at ground
Fixed the camera looking at the ground on load, it was caused by the player moving the mouse while loading the scene. Which resulted in usually looking at the floor.

A small delay has instead been introduced before mouse control is allowed while the scene loads.

Meant to start on pickups but instead pissed about with torch