Added many things
Created placeholder objects for health packs, batteries and keys. They all spawn on the plinth for testing.

Added script for the pickups, it checks the item and applies the affects. there is currently just 1 float that can be changed for each item.

Fixed problem with health bar in HUD as it always read 100 regardless of what the current health actually was.

Responsibility for destroying pickups now lies with the pickups script instead of the object spawner.

Object spawner modified to ensure the item is null before it spawns a replacement due to the new pickup script

Player attributes modified, new methods added for the pickup item effects

Enemy script added to make spikes cause damage to player.

Key logic is yet to be implemented