Added key system
Created Player Inventory script to hold keys and the like

Modified Object spawner to assign a name to keys for doors

Modified Pickups script to add keys to players inventory

Modified Door Interaction script to add a locked or inoperable status to doors, locked doors need keys.

inoperable in door script is just to allow the player to interact with the door but be unable to open it, like a rattle of the handle.

Added a second door for testing, both start as locked. Player starts with red key already in their inventory.