Working on HUD
Added a panel to the UI which displays when the player presses 'F'

panel has 4 slots and a text input for currently carried keys

The inventory sort of works but is not linked properly to the HUD. The background processes seem to work at least.

Items are now collected instead of used on pickup but I haven't implemented a system for using the items although the functions are already written.

I will try again later on, for now it is bed time.