Tidy 2 and atmosphere
Removed colliders on prefabs that were not needed, changed lighting to make the level darker.

I initially removed the directional light and turned the global lighting right down. But that will be a pain in the arse to work in.

Readded the directional light and turned global light back up a bit.

Will control these through script on scene loading.

Hud still a todo