3D compass added
Made a basic 3D compass model and added it to the project, it is currently attached to the player but is inactive. This was for testing purposes.

Edited the compass script to use a module rather than an if statement and renamed a couple of functions for clarity. If you don't like it, I'll change it back and put the needle rotation back in

Compass is just using basic materials for now, I'll make a texture for it later.

Added a needle rotation function to the script.