Glowstick working
The player can now collect and throw glowsticks if they are in an inventory slot. Currently after having picked up a glowstick and thrown it if the player tried to use the same slot again without picking up another glowstick a null reference error occurs. I am aware. I will fix this.

There are 2 versions of the glowstick, 1 is purely for letting the player collect it, the other is for when the player wants to use it.

I spent hours getting this to work.
It works through an object spawner attached to the player.

Colours can be changed in the inspector for the prefab if you fancy a different colour.

Not impressed with the throw force, weak bastard. I have tried with extremely hight numbers and it makes little to no difference, I will work out why  but now I'm starving.